[update] 完成cmap表数据读取，剩余Get方法定义

+++ opentype/src/main/java/am/util/opentype/FileOpenTypeReader.java
@@ -68,7 +68,7 @@ public class FileOpenTypeReader implements OpenTypeReader {
     }
 
     @Override
-    public short readShort() throws IOException {
+    public int readShort() throws IOException {
         return mFile.readShort();
     }
